requirements.txt
chromadb==0.5.3
requests==2.32.3
torch==2.3.1
transformers==4.41.2
pypdf2==3.0.1
sentence_transformers==3.0.1
python-dotenv==1.0.1
django==4.2.13
langchain==0.2.6
langchain-huggingface==0.0.3
huggingface-hub==0.23.4
unstructured==0.14.9
markdown==3.6
psutil==6.0.0
gradio==4.37.2
accelerate==0.31.0
pymongo==4.8.0